moved compute_delay_and_sum_values to delay_and_sum.py
