+++ image_reconstruction/reconstruction_utils/beamforming/__init__.py
@@ -5,3 +5,4 @@ SPDX-License-Identifier: MIT
 
 from .back_projection import back_projection
 from .delay_multiply_and_sum import delay_multiply_and_sum
+from .delay_and_sum import compute_delay_and_sum_values
